read the description pranay (updated rotation amount for climber)
So we added logic to account for if the robot is hit and the position of climber is changed. If climber position is changed, we change how far the motor should move the climber to get it to max position (less than the constant if the climber position is changed).