Added climb mode as command to run when right trigger is pressed.
Notes:

When right trigger is pressed to climb, it will put amp and spivit at maximum angle and then start climbing. The command was made and in robot container, we made it run the command when right trigger is pressed to more than 0.07. Zain did most of the work, just saying.

Aiden approved the climb mode command.